Fix warnings
Former-commit-id: 6240a7e478f124c62a1332a48fc91da7375e83d9
Former-commit-id: 1816d46302ed67e913c9aaf8dfb66523d505c357

+++ framework/src/minko/component/Picking.cpp
@@ -274,8 +274,8 @@ Picking::removedHandler(NodePtr node, NodePtr target, NodePtr parent)
 void
 Picking::renderingBegin(RendererPtr renderer)
 {
-	float mouseX = _mouse->x();
-	float mouseY = _mouse->y();
+	float mouseX = (float)_mouse->x();
+	float mouseY = (float)_mouse->y();
 	
 	auto perspectiveCamera	= _camera->component<component::PerspectiveCamera>();
 	auto projection			= math::Matrix4x4::create()->perspective(perspectiveCamera->fieldOfView(), perspectiveCamera->aspectRatio(), perspectiveCamera->zNear(), perspectiveCamera->zFar());
@@ -286,7 +286,6 @@ Picking::renderingBegin(RendererPtr renderer)
 	pickingProjectionData[6] = mouseY / _context->viewportHeight() * 2.f;
 
 	_pickingProjection->initialize(pickingProjectionData);
-
 }
 
 void
